+++ src/plugin/PluginEditor.h
@@ -40,6 +40,7 @@ private:
     juce::Slider bridgeIntonationSlider;
     juce::Slider fretPressureSlider;
     juce::Slider aftertouchBendSlider;
+    juce::Slider neckSlideSlider;
     juce::Slider fingerNoiseSlider;
     juce::Slider ampFeedbackSlider;
     juce::Slider vibratoSpeedSlider;
@@ -62,6 +63,7 @@ private:
     juce::Label bridgeIntonationLabel;
     juce::Label fretPressureLabel;
     juce::Label aftertouchBendLabel;
+    juce::Label neckSlideLabel;
     juce::Label lookaheadLabel;
     juce::Label fingerNoiseLabel;
     juce::Label ampFeedbackLabel;
@@ -94,6 +96,7 @@ private:
     juce::TextButton bridgeIntonationInfoButton;
     juce::TextButton fretPressureInfoButton;
     juce::TextButton aftertouchBendInfoButton;
+    juce::TextButton neckSlideInfoButton;
     juce::TextButton lookaheadInfoButton;
     juce::TextButton fingerNoiseInfoButton;
     juce::TextButton ampFeedbackInfoButton;
@@ -128,6 +131,7 @@ private:
     std::unique_ptr<SliderAttachment> bridgeIntonationAttachment;
     std::unique_ptr<SliderAttachment> fretPressureAttachment;
     std::unique_ptr<SliderAttachment> aftertouchBendAttachment;
+    std::unique_ptr<SliderAttachment> neckSlideAttachment;
     std::unique_ptr<ComboBoxAttachment> lookaheadAttachment;
     std::unique_ptr<SliderAttachment> fingerNoiseAttachment;
     std::unique_ptr<SliderAttachment> ampFeedbackAttachment;